move signup feature to member controller
