+++ src/main/java/com/neoga/boltacution/memberstore/member/dto/SignupDto.java
@@ -1,4 +1,4 @@
-package com.neoga.boltacution.security.dto;
+package com.neoga.boltacution.memberstore.member.dto;
 
 import lombok.*;
